DRYout operations with the @ field name pattern
