+++ src/Aggregate/Operations/GroupBy.php
@@ -2,23 +2,10 @@
 
 namespace Ehann\RediSearch\Aggregate\Operations;
 
-class GroupBy implements OperationInterface
+class GroupBy extends AbstractFieldNameOperation
 {
-    public $fieldNames = [];
-
     public function __construct(array $fieldNames)
     {
-        $this->fieldNames = $fieldNames;
-    }
-
-    public function toArray(): array
-    {
-        $count = count($this->fieldNames);
-        return $count > 0 ? array_merge([
-            "GROUPBY",
-            $count
-        ], array_map(function ($fieldName) {
-            return "@$fieldName";
-        }, $this->fieldNames)) : [];
+        parent::__construct('GROUPBY', $fieldNames);
     }
 }
