+++ src/Aggregate/Operations/SortBy.php
@@ -2,23 +2,10 @@
 
 namespace Ehann\RediSearch\Aggregate\Operations;
 
-class SortBy implements OperationInterface
+class SortBy extends AbstractFieldNameOperation
 {
-    public $fieldNames = [];
-
     public function __construct(array $fieldNames)
     {
-        $this->fieldNames = $fieldNames;
-    }
-
-    public function toArray(): array
-    {
-        $count = count($this->fieldNames);
-        return $count > 0 ? array_merge(
-            ["SORTBY", $count],
-            implode(' ', array_map(function ($fieldName) {
-                return "@$fieldName";
-            }, $this->fieldNames))
-        ) : [];
+        parent::__construct('SORTBY', $fieldNames);
     }
 }
